pyproject.toml
[tool.poetry]
name = "globa"
version = "0.1.0"
description = ""
authors = ["DL <dmitriy.luzanovsky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.0.6"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
django-config-models = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
